chore(stock): remove the simulated stock market
Refs #544
