+++ .github/PRIVACY.md
@@ -10,14 +10,14 @@ This repository is also published as open source. If you run your own copy, you 
 
 The bot only receives what Discord delivers for the servers it has been invited to and for direct messages sent to it.
 
-| Data                                                                                                         | Where it comes from                                                                                                                                 | Why                                                                                                                                                                                     |
-| ------------------------------------------------------------------------------------------------------------ | --------------------------------------------------------------------------------------------------------------------------------------------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| Message content, author name and ID, channel name and ID, timestamps, attachment and sticker URLs            | Messages from humans in channels the bot can see, direct messages to the bot, and the bot's own replies. Messages from other bots are not recorded. | Producing AI replies that follow the conversation, summarising recent chat, expanding links, and the message-triggered features.                                                        |
-| A per-user long-term memory written in plain text                                                            | Derived by the AI from your own conversations with the bot                                                                                          | Letting the bot remember your preferences across conversations. It is scoped by where it was learned: something told in one server is not surfaced in another.                          |
-| A per-server memory                                                                                          | Derived from public channels only                                                                                                                   | Community context for replies in that server. Content that a server's `@everyone` cannot read never feeds it.                                                                           |
-| Discord user ID, display name, avatar URL, and game state (balances, transactions, positions, match history) | Your use of the economy, casino and stock commands                                                                                                  | Running those features. The currency is virtual, has no monetary value, and no payment information is ever collected.                                                                   |
-| Feature usage records containing numeric IDs only                                                            | Each slash command invocation and each AI reply                                                                                                     | Knowing which features are used. No names, no message content and no command arguments are stored in these records.                                                                     |
-| Runtime diagnostic logs                                                                                      | The bot process                                                                                                                                     | Debugging. These may contain excerpts of message content and stay on the operator's machine; telemetry is explicitly disabled and nothing is sent to an external observability service. |
+| Data                                                                                              | Where it comes from                                                                                                                                 | Why                                                                                                                                                                                     |
+| ------------------------------------------------------------------------------------------------- | --------------------------------------------------------------------------------------------------------------------------------------------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| Message content, author name and ID, channel name and ID, timestamps, attachment and sticker URLs | Messages from humans in channels the bot can see, direct messages to the bot, and the bot's own replies. Messages from other bots are not recorded. | Producing AI replies that follow the conversation, summarising recent chat, expanding links, and the message-triggered features.                                                        |
+| A per-user long-term memory written in plain text                                                 | Derived by the AI from your own conversations with the bot                                                                                          | Letting the bot remember your preferences across conversations. It is scoped by where it was learned: something told in one server is not surfaced in another.                          |
+| A per-server memory                                                                               | Derived from public channels only                                                                                                                   | Community context for replies in that server. Content that a server's `@everyone` cannot read never feeds it.                                                                           |
+| Discord user ID, display name, avatar URL, and game state (balances, transactions, match history) | Your use of the economy and casino commands                                                                                                         | Running those features. The currency is virtual, has no monetary value, and no payment information is ever collected.                                                                   |
+| Feature usage records containing numeric IDs only                                                 | Each slash command invocation and each AI reply                                                                                                     | Knowing which features are used. No names, no message content and no command arguments are stored in these records.                                                                     |
+| Runtime diagnostic logs                                                                           | The bot process                                                                                                                                     | Debugging. These may contain excerpts of message content and stay on the operator's machine; telemetry is explicitly disabled and nothing is sent to an external observability service. |
 
 Everything above is stored in local SQLite databases and local files on the operator's machine.
