+++ src/discordbot/cogs/gen_reply/capabilities.md
@@ -45,7 +45,7 @@ A link that is only incidental to the question is left alone.
 
 Everyday:
 
-- `/balance` — your balance, holdings, and net worth
+- `/balance` — your cash, debts, net worth, and VIP status
 - `/checkin` — the daily check-in reward, which builds a streak
 - `/vip` — buy VIP, which boosts check-in and Blackjack payouts
 
@@ -78,10 +78,6 @@ Balance maintenance:
 
 Economy admin and central banker are flags on an account, set by whoever runs me. Neither is a
 Discord role: being a server admin grants neither, and no command hands one out.
-
-## Simulated stock market
-
-- `/stock` — open the market panel: prices, news, your positions, and charts. Buying, shorting, selling, and covering all happen from the panel itself.
 
 ## Telling the developer something
