--- src/discordbot/services/stock/market.py
@@ -1,231 +1,0 @@
-"""Pure market helpers for the simulated stock market."""
-
-from random import Random
-from datetime import UTC, datetime, timedelta
-
-from discordbot.typings.stock import STOCK_TICK_SECONDS, MAX_TICKS_PER_INTERACTION
-from discordbot.utils.timezone import TAIWAN_TIMEZONE, as_taipei
-
-NEWS_SENTIMENT_DECAY_BPS = 20
-NEWS_SENTIMENT_DECAY_SECONDS = 60 * 60
-NEWS_SENTIMENT_LIMIT_BPS = 300
-PRESSURE_LIMIT_BPS = 60
-
-# Anti-inflation / realism guardrails for the price formula. A 5-minute tick at
-# real-market magnitudes moves fractions of a percent, so the per-company random
-# volatility is scaled down toward that range, every tick is hard-capped no matter
-# how aggressive the per-company knob is, and each Asia/Taipei trading day is bounded
-# to a Taiwan-style price limit measured against the previous close. Reducing per-tick
-# swing and capping the daily move shrinks the volatility-harvesting that mints money.
-# MARKET_VOLATILITY_SCALE_BPS was calibrated by offline simulation: at 800 the most
-# aggressive production profile lands near 70% annualized volatility with the daily
-# limit binding only a few percent of days, and calmer profiles scale down from there.
-MARKET_VOLATILITY_SCALE_BPS = 800
-GLOBAL_MAX_TICK_CHANGE_BPS = 200
-DAILY_PRICE_LIMIT_BPS = 1_000
-
-
-def format_price(price_cents: int) -> str:
-    """Formats a cent-denominated stock price."""
-    return f"{price_cents // 100:,}.{price_cents % 100:02d}"
-
-
-def clamp_bps(value: int, lower: int, upper: int) -> int:
-    """Clamps basis-point values."""
-    return max(lower, min(upper, value))
-
-
-def tick_boundary(dt: datetime) -> datetime:
-    """Returns the tick-boundary timestamp at or before `dt`."""
-    taipei_dt = as_taipei(dt=dt)
-    seconds = int(taipei_dt.timestamp())
-    boundary = seconds - (seconds % STOCK_TICK_SECONDS)
-    return datetime.fromtimestamp(boundary, tz=UTC).astimezone(tz=TAIWAN_TIMEZONE)
-
-
-def decay_news_sentiment(sentiment_bps: int, elapsed_seconds: int) -> int:
-    """Applies linear time-based news sentiment decay and clamps the result.
-
-    Ambient context for the news-generation prompt only. Pricing applies each
-    news row as a one-shot impulse at its own firing tick boundary, never as a
-    decayed drift over the ticks after it.
-    """
-    clamped = clamp_bps(
-        value=sentiment_bps, lower=-NEWS_SENTIMENT_LIMIT_BPS, upper=NEWS_SENTIMENT_LIMIT_BPS
-    )
-    decay_bps = max(elapsed_seconds, 0) * NEWS_SENTIMENT_DECAY_BPS // NEWS_SENTIMENT_DECAY_SECONDS
-    remaining = max(abs(clamped) - decay_bps, 0)
-    return remaining if clamped >= 0 else -remaining
-
-
-def pressure_from_order_flow(net_shares: float, liquidity_shares: int) -> int:
-    """Converts decayed net order flow into bounded price pressure."""
-    if liquidity_shares <= 0 or net_shares == 0:
-        return 0
-    return clamp_bps(
-        value=round(net_shares * PRESSURE_LIMIT_BPS / liquidity_shares),
-        lower=-PRESSURE_LIMIT_BPS,
-        upper=PRESSURE_LIMIT_BPS,
-    )
-
-
-def order_impact_bps(shares: int, liquidity_shares: int, max_impact_bps: int) -> int:
-    """Converts an order size into bounded execution impact."""
-    if shares <= 0 or liquidity_shares <= 0 or max_impact_bps <= 0:
-        return 0
-    raw_impact, remainder = divmod(shares * max_impact_bps, liquidity_shares)
-    if remainder * 2 >= liquidity_shares:
-        raw_impact += 1
-    return clamp_bps(value=raw_impact, lower=0, upper=max_impact_bps)
-
-
-def execution_price_cents(
-    reference_price_cents: int,
-    shares: int,
-    liquidity_shares: int,
-    max_impact_bps: int,
-    is_buy: bool,
-) -> int:
-    """Returns a side-adjusted execution price for one order leg."""
-    reference_price = max(reference_price_cents, 1)
-    impact_bps = order_impact_bps(
-        shares=shares, liquidity_shares=liquidity_shares, max_impact_bps=max_impact_bps
-    )
-    if is_buy:
-        return max(reference_price * (10_000 + impact_bps) + 9_999, 1) // 10_000
-    return max(reference_price * max(10_000 - impact_bps, 1) // 10_000, 1)
-
-
-def mean_reversion_bps(
-    previous_price_cents: int, fair_value_cents: int, mean_reversion_strength_bps: int
-) -> int:
-    """Returns the fair-value pull for the next tick, scaled by the reversion strength."""
-    if previous_price_cents <= 0 or fair_value_cents <= 0 or mean_reversion_strength_bps <= 0:
-        return 0
-    fair_value_gap_bps = (fair_value_cents - previous_price_cents) * 10_000 // previous_price_cents
-    return fair_value_gap_bps * mean_reversion_strength_bps // 10_000
-
-
-def effective_volatility_width_bps(base_volatility_bps: int, volatility_amplifier_bps: int) -> int:
-    """Returns the scaled per-tick random volatility half-width in basis points."""
-    raw_width = base_volatility_bps * volatility_amplifier_bps // 100
-    return max(raw_width * MARKET_VOLATILITY_SCALE_BPS // 10_000, 0)
-
-
-def apply_daily_price_limit(price_cents: int, previous_close_cents: int, limit_bps: int) -> int:
-    """Clamps a tick price to a Taiwan-style daily band around the previous close."""
-    if previous_close_cents <= 0 or limit_bps <= 0:
-        return max(price_cents, 1)
-    upper = previous_close_cents * (10_000 + limit_bps) // 10_000
-    lower = max(previous_close_cents * (10_000 - limit_bps) // 10_000, 1)
-    return max(min(price_cents, upper), lower)
-
-
-def calculate_next_price_cents(  # noqa: PLR0913 -- pure price formula takes every market factor explicitly
-    previous_price_cents: int,
-    news_sentiment_bps: int,
-    pressure_bps: int,
-    base_volatility_bps: int,
-    volatility_amplifier_bps: int,
-    fair_value_cents: int,
-    mean_reversion_strength_bps: int,
-    max_tick_change_bps: int,
-    rng: Random,
-) -> int:
-    """Calculates the next price using deterministic inputs plus seeded randomness."""
-    volatility_width = effective_volatility_width_bps(
-        base_volatility_bps=base_volatility_bps, volatility_amplifier_bps=volatility_amplifier_bps
-    )
-    random_bps = rng.randint(-volatility_width, volatility_width) if volatility_width else 0
-    raw_change_bps = (
-        random_bps
-        + clamp_bps(
-            value=news_sentiment_bps,
-            lower=-NEWS_SENTIMENT_LIMIT_BPS,
-            upper=NEWS_SENTIMENT_LIMIT_BPS,
-        )
-        + clamp_bps(value=pressure_bps, lower=-PRESSURE_LIMIT_BPS, upper=PRESSURE_LIMIT_BPS)
-        + mean_reversion_bps(
-            previous_price_cents=previous_price_cents,
-            fair_value_cents=fair_value_cents,
-            mean_reversion_strength_bps=mean_reversion_strength_bps,
-        )
-    )
-    change_limit = max(min(max_tick_change_bps, GLOBAL_MAX_TICK_CHANGE_BPS), 1)
-    change_bps = clamp_bps(value=raw_change_bps, lower=-change_limit, upper=change_limit)
-    next_price = previous_price_cents * (10_000 + change_bps) // 10_000
-    return max(next_price, 1)
-
-
-def _tick_boundaries_between(
-    latest_boundary: datetime, current_boundary: datetime
-) -> list[datetime]:
-    """Returns all tick boundaries between two normalized endpoints."""
-    boundaries: list[datetime] = []
-    boundary = latest_boundary + timedelta(seconds=STOCK_TICK_SECONDS)
-    while boundary <= current_boundary:
-        boundaries.append(boundary)
-        boundary += timedelta(seconds=STOCK_TICK_SECONDS)
-    return boundaries
-
-
-def _compressed_required_boundaries(
-    latest_boundary: datetime, current_boundary: datetime, boundaries: list[datetime]
-) -> set[datetime]:
-    """Returns boundaries that must survive compression for day rollover correctness."""
-    selected = {current_boundary}
-    boundary_set = set(boundaries)
-    previous_boundary = latest_boundary
-    for boundary in boundaries:
-        if as_taipei(dt=boundary).date() != as_taipei(dt=previous_boundary).date():
-            selected.add(boundary)
-            if previous_boundary in boundary_set:
-                selected.add(previous_boundary)
-        previous_boundary = boundary
-    return selected
-
-
-def _fill_compressed_boundaries(
-    boundaries: list[datetime], selected: set[datetime]
-) -> tuple[datetime, ...]:
-    """Fills a compressed boundary set with even sampling plus recent fallbacks."""
-    total_steps = len(boundaries)
-    for index in range(MAX_TICKS_PER_INTERACTION):
-        source_index = (index + 1) * total_steps // MAX_TICKS_PER_INTERACTION - 1
-        selected.add(boundaries[source_index])
-        if len(selected) >= MAX_TICKS_PER_INTERACTION:
-            break
-    for boundary in reversed(boundaries):
-        if len(selected) >= MAX_TICKS_PER_INTERACTION:
-            break
-        selected.add(boundary)
-    return tuple(sorted(selected))
-
-
-def tick_boundaries_to_apply(latest_tick_at: datetime, now: datetime) -> tuple[datetime, ...]:
-    """Returns tick boundaries that should be materialized for a lazy interaction.
-
-    A backlog longer than `MAX_TICKS_PER_INTERACTION` is compressed to that many
-    boundaries. Each Asia/Taipei day rollover keeps both the boundary that opens
-    the new day and the one before it, so the caller still sees every previous
-    close and day open; a backlog holding more rollovers than the cap keeps only
-    the most recent ones.
-    """
-    latest_boundary = tick_boundary(dt=latest_tick_at)
-    current_boundary = tick_boundary(dt=now)
-    if current_boundary <= latest_boundary:
-        return ()
-
-    boundaries = _tick_boundaries_between(
-        latest_boundary=latest_boundary, current_boundary=current_boundary
-    )
-
-    if len(boundaries) <= MAX_TICKS_PER_INTERACTION:
-        return tuple(boundaries)
-
-    selected = _compressed_required_boundaries(
-        latest_boundary=latest_boundary, current_boundary=current_boundary, boundaries=boundaries
-    )
-    if len(selected) >= MAX_TICKS_PER_INTERACTION:
-        return tuple(sorted(selected)[-MAX_TICKS_PER_INTERACTION:])
-    return _fill_compressed_boundaries(boundaries=boundaries, selected=selected)
