--- src/discordbot/typings/stock.py
@@ -1,464 +1,0 @@
-"""Shared vocabulary for the stock market simulation.
-
-Two money units live here and only the field name tells them apart: a `*_cents` field is
-cent-denominated, and every other money field is whole wallet cash, already rounded out of
-cents by `cash_ceil` / `cash_floor` before it was stored.
-"""
-
-from enum import StrEnum
-from typing import Self, Final
-from datetime import datetime
-
-from pydantic import Field, BaseModel, ConfigDict, model_validator
-
-STOCK_TICK_SECONDS: Final[int] = 5 * 60
-MAX_TICKS_PER_INTERACTION: Final[int] = 12 * 24
-STOCK_HISTORY_DAYS: Final[int] = 7
-STOCK_ACTION_TIMEOUT_SECONDS: Final[int] = 180
-STOCK_NEWS_CADENCE_HOURS: Final[int] = 4
-STOCK_INDIVIDUAL_OWNERSHIP_CAP_BPS: Final[int] = 4_900
-STOCK_BPS_DENOMINATOR: Final[int] = 10_000
-
-
-class StockAction(StrEnum):
-    """User-facing stock operation families."""
-
-    BUY = "buy"
-    SHORT = "short"
-
-
-class StockOperationStatus(StrEnum):
-    """Lifecycle states for a cross-database stock operation."""
-
-    PENDING = "pending"
-    WALLET_APPLIED = "wallet_applied"
-    APPLIED = "applied"
-    FAILED = "failed"
-    RECONCILE_REQUIRED = "reconcile_required"
-
-
-class StockTradeLegType(StrEnum):
-    """Atomic legs generated by one stock operation."""
-
-    OPEN_LONG = "open_long"
-    SELL_LONG = "sell_long"
-    OPEN_SHORT = "open_short"
-    COVER_SHORT = "cover_short"
-
-
-class StockProfileView(BaseModel):
-    """Read-only stock profile and latest quote."""
-
-    model_config = ConfigDict(frozen=True)
-
-    symbol: str = Field(..., description="Virtual company ticker symbol.")
-    name: str = Field(..., description="Display name of the virtual company.")
-    category: str = Field(..., description="Market category the company belongs to.")
-    price_cents: int = Field(..., description="Latest quote price in cents.")
-    previous_close_price_cents: int = Field(
-        ..., description="Previous trading-day close price in cents."
-    )
-    day_open_price_cents: int = Field(..., description="Current trading-day open price in cents.")
-    total_shares: int = Field(..., description="Total issued share count.")
-    float_shares: int = Field(..., description="Tradable float share count.")
-    base_volatility_bps: int = Field(
-        ..., description="Baseline per-tick volatility in basis points."
-    )
-    volatility_amplifier_bps: int = Field(
-        ..., description="Additional volatility amplifier in basis points."
-    )
-    liquidity_shares: int = Field(
-        ..., description="Liquidity depth in shares used for order-size slippage."
-    )
-    fair_value_cents: int = Field(..., description="Mean-reversion fair value anchor in cents.")
-    mean_reversion_bps: int = Field(
-        ..., description="Per-tick mean-reversion pull toward fair value in basis points."
-    )
-    max_tick_change_bps: int = Field(
-        ..., description="Cap on price change per tick in basis points."
-    )
-    news_cadence_hours: int = Field(
-        ..., description="Minimum hours between news refreshes for this symbol."
-    )
-    updated_at: datetime = Field(..., description="Timestamp of the latest profile update.")
-
-
-class StockProfileUpsert(BaseModel):
-    """DB-owned stock profile payload for maintenance scripts."""
-
-    model_config = ConfigDict(frozen=True)
-
-    symbol: str = Field(
-        min_length=1,
-        max_length=16,
-        description="Virtual company ticker symbol.",
-        examples=["ACME"],
-    )
-    name: str = Field(
-        min_length=1,
-        max_length=128,
-        description="Display name of the virtual company.",
-        examples=["Acme Corp"],
-    )
-    category: str = Field(
-        min_length=1,
-        max_length=64,
-        description="Market category the company belongs to.",
-        examples=["tech"],
-    )
-    price_cents: int = Field(ge=1, description="Initial quote price in cents.", examples=[10000])
-    total_shares: int = Field(ge=1, description="Total issued share count.", examples=[1000000])
-    float_shares: int = Field(ge=0, description="Tradable float share count.", examples=[500000])
-    base_volatility_bps: int = Field(
-        ge=0, description="Baseline per-tick volatility in basis points.", examples=[50]
-    )
-    volatility_amplifier_bps: int = Field(
-        ge=0, description="Additional volatility amplifier in basis points.", examples=[20]
-    )
-    liquidity_shares: int = Field(
-        ge=1,
-        description="Liquidity depth in shares used for order-size slippage.",
-        examples=[10000],
-    )
-    fair_value_cents: int = Field(
-        ge=1, description="Mean-reversion fair value anchor in cents.", examples=[10000]
-    )
-    mean_reversion_bps: int = Field(
-        ge=0,
-        description="Per-tick mean-reversion pull toward fair value in basis points.",
-        examples=[10],
-    )
-    max_tick_change_bps: int = Field(
-        ge=1, description="Cap on price change per tick in basis points.", examples=[300]
-    )
-    news_cadence_hours: int = Field(
-        ge=1, description="Minimum hours between news refreshes for this symbol.", examples=[4]
-    )
-
-    @model_validator(mode="after")
-    def validate_share_structure(self) -> Self:
-        """Validates share counts that depend on each other."""
-        if self.float_shares > self.total_shares:
-            msg = "float_shares cannot exceed total_shares"
-            raise ValueError(msg)
-        return self
-
-
-class StockPositionView(BaseModel):
-    """Read-only per-user stock position."""
-
-    model_config = ConfigDict(frozen=True)
-
-    symbol: str = Field(..., description="Virtual company ticker symbol.")
-    user_id: int = Field(..., description="Discord user ID owning the position.")
-    user_name: str = Field(default="", description="Stored display name of the position owner.")
-    long_shares: int = Field(default=0, description="Number of shares held long.")
-    long_cost_basis: int = Field(
-        default=0, description="Aggregate cost basis of the long position, in cash."
-    )
-    short_shares: int = Field(default=0, description="Number of shares held short.")
-    short_entry_value: int = Field(
-        default=0, description="Aggregate entry value of the short position, in cash."
-    )
-    short_collateral: int = Field(
-        default=0, description="Collateral reserved against the short position, in cash."
-    )
-    realized_pnl: int = Field(
-        default=0, description="Realized profit and loss for this position, in cash."
-    )
-
-
-class StockParticipantPositionView(BaseModel):
-    """Public stock-level position summary for one participant."""
-
-    model_config = ConfigDict(frozen=True)
-
-    user_id: int = Field(..., description="Discord user ID of the participant.")
-    user_name: str = Field(..., description="Stored display name of the participant.")
-    long_shares: int = Field(..., description="Number of shares the participant holds long.")
-    short_shares: int = Field(..., description="Number of shares the participant holds short.")
-    realized_pnl: int = Field(
-        ..., description="Realized profit and loss for the participant, in cash."
-    )
-
-
-class StockTradeLegView(BaseModel):
-    """Read-only stock trade leg for UI and reconciliation."""
-
-    model_config = ConfigDict(frozen=True)
-
-    operation_id: str = Field(..., description="Parent stock operation identifier.")
-    leg_order: int = Field(..., description="Ordering index of this leg within the operation.")
-    symbol: str = Field(..., description="Virtual company ticker symbol.")
-    user_id: int = Field(..., description="Discord user ID the leg belongs to.")
-    user_name: str = Field(default="", description="Stored display name of the leg owner.")
-    leg_type: StockTradeLegType = Field(..., description="Type of atomic trade leg.")
-    shares: int = Field(..., description="Share quantity executed in this leg.")
-    price_cents: int = Field(..., description="Per-leg execution price in cents.")
-    wallet_delta: int = Field(..., description="Wallet balance change applied by this leg.")
-    basis_delta: int = Field(..., description="Cost-basis change applied by this leg, in cash.")
-    collateral_delta: int = Field(
-        ..., description="Short collateral change applied by this leg, in cash."
-    )
-    realized_pnl_delta: int = Field(
-        ..., description="Realized profit and loss change applied by this leg, in cash."
-    )
-    created_at: datetime = Field(..., description="Timestamp the leg was created.")
-
-
-class StockNewsView(BaseModel):
-    """Read-only stock news item."""
-
-    model_config = ConfigDict(frozen=True)
-
-    symbol: str = Field(..., description="Virtual company ticker symbol.")
-    headline: str = Field(..., description="News headline text.")
-    sentiment_bps: int = Field(..., description="News sentiment impulse in basis points.")
-    source: str = Field(
-        default="template", description="Origin of the news item: `template` or `ai`."
-    )
-    model: str = Field(default="", description="Model identifier that generated the news, if any.")
-    expires_at: datetime | None = Field(
-        default=None, description="Timestamp after which the news is stale, if set."
-    )
-    created_at: datetime = Field(..., description="Timestamp the news item was created.")
-
-
-class StockGeneratedNews(BaseModel):
-    """Generated stock news payload before persistence."""
-
-    model_config = ConfigDict(frozen=True)
-
-    headline: str = Field(..., description="Generated news headline text.")
-    sentiment_bps: int = Field(
-        ..., description="Generated news sentiment impulse in basis points."
-    )
-    source: str = Field(..., description="Origin of the generated news: `template` or `ai`.")
-    model: str = Field(default="", description="Model identifier that generated the news, if any.")
-
-
-class StockNewsGenerationContext(BaseModel):
-    """Market context passed to a stock news generator."""
-
-    model_config = ConfigDict(frozen=True)
-
-    profile: StockProfileView = Field(
-        ..., description="Stock profile and latest quote for context."
-    )
-    change_cents: int = Field(..., description="Daily price change in cents.")
-    change_bps: int = Field(..., description="Daily price change in basis points.")
-    pressure_bps: int = Field(
-        ..., description="Recent decayed order-flow pressure in basis points."
-    )
-    buy_side_shares: int = Field(..., description="Recent buy-side order flow in shares.")
-    sell_side_shares: int = Field(..., description="Recent sell-side order flow in shares.")
-    net_order_shares: int = Field(..., description="Net order flow in shares.")
-    recent_news_sentiment_bps: int = Field(
-        ..., description="Existing decayed news sentiment in basis points."
-    )
-    latest_news_headline: str = Field(default="", description="Most recent news headline, if any.")
-    latest_news_sentiment_bps: int = Field(
-        default=0, description="Sentiment of the most recent news in basis points."
-    )
-    lookback_hours: int = Field(
-        ..., description="Lookback window in hours used to compute the context."
-    )
-
-
-class StockPriceTickView(BaseModel):
-    """Read-only stock price tick."""
-
-    model_config = ConfigDict(frozen=True)
-
-    symbol: str = Field(..., description="Virtual company ticker symbol.")
-    price_cents: int = Field(..., description="Quote price at this tick in cents.")
-    created_at: datetime = Field(..., description="Timestamp of the price tick.")
-
-
-class StockMarketQuote(BaseModel):
-    """One row in the market list."""
-
-    model_config = ConfigDict(frozen=True)
-
-    profile: StockProfileView = Field(..., description="Stock profile and latest quote.")
-    change_cents: int = Field(..., description="Daily price change in cents.")
-    change_bps: int = Field(..., description="Daily price change in basis points.")
-    pressure_bps: int = Field(..., description="Recent order-flow pressure in basis points.")
-
-
-class StockDetailViewData(BaseModel):
-    """Full personal detail view for one stock."""
-
-    model_config = ConfigDict(frozen=True)
-
-    quote: StockMarketQuote = Field(..., description="Market quote for the selected stock.")
-    balance: int = Field(..., description="Viewing user's wallet cash balance.")
-    position: StockPositionView = Field(
-        ..., description="Viewing user's current position in the stock."
-    )
-    recent_trades: tuple[StockTradeLegView, ...] = Field(
-        ..., description="Recent trade legs for the viewing user."
-    )
-    public_positions: tuple[StockParticipantPositionView, ...] = Field(
-        default=(), description="Public position summaries for all participants."
-    )
-    news: tuple[StockNewsView, ...] = Field(..., description="Recent news items for the stock.")
-    ticks: tuple[StockPriceTickView, ...] = Field(
-        ..., description="Recent price ticks for the 7D chart."
-    )
-
-
-class StockPortfolioHolding(BaseModel):
-    """One current stock position row for a user's portfolio."""
-
-    model_config = ConfigDict(frozen=True)
-
-    symbol: str = Field(..., description="Virtual company ticker symbol.")
-    name: str = Field(..., description="Display name of the virtual company.")
-    price_cents: int = Field(..., description="Latest quote price in cents.")
-    long_shares: int = Field(..., description="Number of shares held long.")
-    long_cost_basis: int = Field(
-        ..., description="Aggregate cost basis of the long position, in cash."
-    )
-    long_market_value: int = Field(
-        ..., description="Current market value of the long position, in cash."
-    )
-    short_shares: int = Field(..., description="Number of shares held short.")
-    short_entry_value: int = Field(
-        ..., description="Aggregate entry value of the short position, in cash."
-    )
-    short_collateral: int = Field(
-        ..., description="Collateral reserved against the short position, in cash."
-    )
-    short_cover_cost: int = Field(
-        ..., description="Current cost to cover the short position, in cash."
-    )
-    equity_value: int = Field(..., description="Net equity value of this holding, in cash.")
-    unrealized_pnl: int = Field(
-        ..., description="Unrealized profit and loss for this holding, in cash."
-    )
-    realized_pnl: int = Field(
-        ..., description="Realized profit and loss for this holding, in cash."
-    )
-
-
-class StockPortfolioView(BaseModel):
-    """Current stock portfolio value for one user."""
-
-    model_config = ConfigDict(frozen=True)
-
-    user_id: int = Field(..., description="Discord user ID owning the portfolio.")
-    holdings: tuple[StockPortfolioHolding, ...] = Field(
-        ..., description="Current stock holdings for the user."
-    )
-    equity_value: int = Field(
-        ..., description="Total net equity value across all holdings, in cash."
-    )
-    unrealized_pnl: int = Field(
-        ..., description="Total unrealized profit and loss across holdings, in cash."
-    )
-    realized_pnl: int = Field(
-        ..., description="Total realized profit and loss across holdings, in cash."
-    )
-
-
-class StockSettlementResult(BaseModel):
-    """Result returned by the one stock settlement helper."""
-
-    model_config = ConfigDict(frozen=True)
-
-    success: bool = Field(..., description="Whether the settlement completed successfully.")
-    operation_id: str | None = Field(
-        ..., description="Identifier of the settled operation, if created."
-    )
-    symbol: str = Field(..., description="Virtual company ticker symbol.")
-    requested_action: StockAction = Field(
-        ..., description="Stock operation family requested by the user."
-    )
-    shares: int = Field(..., description="Share quantity settled.")
-    price_cents: int = Field(..., description="Execution price in cents.")
-    wallet_delta: int = Field(..., description="Net wallet balance change from the settlement.")
-    balance_after: int = Field(..., description="Wallet cash balance after settlement.")
-    position: StockPositionView = Field(..., description="Resulting position after settlement.")
-    legs: tuple[StockTradeLegView, ...] = Field(
-        ..., description="Trade legs produced by the settlement."
-    )
-    status: StockOperationStatus | None = Field(
-        default=None, description="Final operation lifecycle status, if known."
-    )
-    error: str = Field(default="", description="Error message when the settlement fails.")
-
-
-class StockReconciliationOperation(BaseModel):
-    """Read-only non-final operation snapshot for manual reconciliation."""
-
-    model_config = ConfigDict(frozen=True)
-
-    operation_id: str = Field(..., description="Stock operation identifier.")
-    status: StockOperationStatus = Field(..., description="Current operation lifecycle status.")
-    user_id: int = Field(..., description="Discord user ID that initiated the operation.")
-    user_name: str = Field(..., description="Stored display name of the operation owner.")
-    symbol: str = Field(..., description="Virtual company ticker symbol.")
-    requested_action: StockAction = Field(
-        ..., description="Stock operation family requested by the user."
-    )
-    failure_reason: str = Field(..., description="Recorded reason the operation did not finalize.")
-    created_at: datetime = Field(..., description="Timestamp the operation was created.")
-    updated_at: datetime = Field(..., description="Timestamp of the latest operation update.")
-    legs: tuple[StockTradeLegView, ...] = Field(
-        ..., description="Trade legs associated with the operation."
-    )
-
-
-class StockSupplyAuditView(BaseModel):
-    """Read-only stock supply and exposure snapshot for maintenance."""
-
-    model_config = ConfigDict(frozen=True)
-
-    symbol: str = Field(..., description="Virtual company ticker symbol.")
-    name: str = Field(..., description="Display name of the virtual company.")
-    price_cents: int = Field(..., description="Latest quote price in cents.")
-    total_shares: int = Field(..., description="Total issued share count.")
-    float_shares: int = Field(..., description="Tradable float share count.")
-    long_shares: int = Field(..., description="Aggregate long exposure in shares.")
-    short_shares: int = Field(..., description="Aggregate short borrow in shares.")
-    available_long_shares: int = Field(..., description="Remaining long capacity in shares.")
-    available_short_shares: int = Field(
-        ..., description="Remaining short borrow capacity in shares."
-    )
-    liquidity_shares: int = Field(
-        ..., description="Liquidity depth in shares used for order-size slippage."
-    )
-    non_final_operations: int = Field(
-        ..., description="Count of operations not yet in a final state."
-    )
-
-
-__all__ = [
-    "MAX_TICKS_PER_INTERACTION",
-    "STOCK_ACTION_TIMEOUT_SECONDS",
-    "STOCK_BPS_DENOMINATOR",
-    "STOCK_HISTORY_DAYS",
-    "STOCK_INDIVIDUAL_OWNERSHIP_CAP_BPS",
-    "STOCK_NEWS_CADENCE_HOURS",
-    "STOCK_TICK_SECONDS",
-    "StockAction",
-    "StockDetailViewData",
-    "StockGeneratedNews",
-    "StockMarketQuote",
-    "StockNewsGenerationContext",
-    "StockNewsView",
-    "StockOperationStatus",
-    "StockParticipantPositionView",
-    "StockPortfolioHolding",
-    "StockPortfolioView",
-    "StockPositionView",
-    "StockPriceTickView",
-    "StockProfileUpsert",
-    "StockProfileView",
-    "StockReconciliationOperation",
-    "StockSettlementResult",
-    "StockSupplyAuditView",
-    "StockTradeLegType",
-    "StockTradeLegView",
-]
