+++ tests/test_number_text.py
@@ -1,6 +1,6 @@
 """Tests for shared readable number presentation helpers."""
 
-from discordbot.utils.number_text import compact_amount, compact_number, share_quantity_text
+from discordbot.utils.number_text import compact_amount, compact_number
 from discordbot.services.economy.presentation import amount_code, currency_text
 
 
@@ -39,13 +39,3 @@ def test_currency_helpers_can_opt_into_compact_amounts() -> None:
 def test_compact_number_matches_amount_formatting() -> None:
     """Generic numeric text keeps the same compact scale behavior."""
     assert compact_number(number=123_456_789) == "1.23億"
-
-
-def test_share_quantity_text_uses_lot_units_without_changing_small_shares() -> None:
-    """Stock share display switches to 張 only after one lot."""
-    assert share_quantity_text(shares=999) == "999股"
-    assert share_quantity_text(shares=1_000) == "1張"
-    assert share_quantity_text(shares=1_234) == "1張 234股"
-    assert share_quantity_text(shares=-1_234) == "-1張 234股"
-    assert share_quantity_text(shares=1_234, signed=True) == "+1張 234股"
-    assert share_quantity_text(shares=10_000_000_000_000) == "100億張"
